mails/requirements-2.txt
flask
sparkpost
Flask-Mail
python-decouple